convert to float tensor as default

+++ model_utils.lua
@@ -6,7 +6,7 @@ function model_utils.transfer_data(x, should_cuda)
     if should_cuda then
         return x:cuda()
     else
-        return x:double()
+        return x:float()
     end
 end
